@@ -8,20 +8,27 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
+import androidx.recyclerview.widget.GridLayoutManager;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
 
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.unipet7.mcommerce.R;
+import com.unipet7.mcommerce.adapters.CartAdapter;
+import com.unipet7.mcommerce.adapters.Cart_Checkout_Adapter;
 import com.unipet7.mcommerce.adapters.PaymentMethodListener;
 import com.unipet7.mcommerce.databinding.ActivityNotificationBinding;
 import com.unipet7.mcommerce.databinding.ActivityPaymentMethodDetailsBinding;
 import com.unipet7.mcommerce.firebase.FireStoreClass;
 import com.unipet7.mcommerce.fragments.FragmentAddressEdit;
 import com.unipet7.mcommerce.fragments.FragmentPaymentMethod;
+import com.unipet7.mcommerce.models.ProductCart;
 import com.unipet7.mcommerce.models.User;
 
 import java.util.ArrayList;
@@ -29,13 +36,39 @@ import java.util.ArrayList;
 public class PaymentMethodDetails extends AppCompatActivity implements PaymentMethodListener {
     private static final int REQUEST_CODE_VOUCHER = 101;
     ActivityPaymentMethodDetailsBinding binding;
+    Cart_Checkout_Adapter adapter;
+    ArrayList<ProductCart> productCarts;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         binding = ActivityPaymentMethodDetailsBinding.inflate(getLayoutInflater());
         setActionBaricon();
         addEvents();
+        loadData();
         setContentView(binding.getRoot());
+    }
+
+    private void loadData() {
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        String currentUserId = new FireStoreClass().getCurrentUID();
+        if (currentUserId != null && !currentUserId.isEmpty()) {
+            db.collection("cart")
+                    .whereEqualTo("userId", currentUserId)
+                    .get()
+                    .addOnSuccessListener(queryDocumentSnapshots -> {
+                        productCarts = new ArrayList<>();
+                        for (QueryDocumentSnapshot document : queryDocumentSnapshots) {
+                            ProductCart productCart = document.toObject(ProductCart.class);
+                            productCarts.add(productCart);
+                        }
+                        adapter = new Cart_Checkout_Adapter(PaymentMethodDetails.this, productCarts);
+                        binding.rclPurchasedProduct.setLayoutManager(new GridLayoutManager(PaymentMethodDetails.this, 1));
+                        binding.rclPurchasedProduct.setAdapter(adapter);
+                    })
+                    .addOnFailureListener(e -> Log.e("Firestore", "Error fetching documents: ", e));
+        } else {
+            Log.e("Firestore", "Current user ID is null or empty");
+        }
     }
 
     private void setActionBaricon() {
@@ -52,17 +85,14 @@ public class PaymentMethodDetails extends AppCompatActivity implements PaymentMe
     }
 
     private void addEvents() {
-        binding.txtChangePaymentMethod.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FragmentPaymentMethod editAdrressFragment = new FragmentPaymentMethod();
-                getSupportFragmentManager()
-                        .beginTransaction()
-                        .add(R.id.containerpayment, editAdrressFragment, "editAddressFragment")
-                        .addToBackStack(null)
-                        .commit();
+        binding.txtChangePaymentMethod.setOnClickListener(v -> {
+            FragmentPaymentMethod editAdrressFragment = new FragmentPaymentMethod();
+            getSupportFragmentManager()
+                    .beginTransaction()
+                    .add(R.id.containerpayment, editAdrressFragment, "editAddressFragment")
+                    .addToBackStack(null)
+                    .commit();
 
-            }
         });
     }
     public void onPaymentMethodSelected(String paymentMethod) {
