@@ -1,27 +1,24 @@
 package com.unipet7.mcommerce.adapters;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.EditText;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.models.Addresses;
 
-import java.util.ArrayList;
+import java.util.List;
 
 public class AddressAdapter extends BaseAdapter {
     Context context;
-    ArrayList<Addresses> addresses;
+    List<Addresses> addresses;
     int item_layout;
 
-    public AddressAdapter(Context context, ArrayList<Addresses> addresses, int item_layout) {
+    public AddressAdapter(Context context, List<Addresses> addresses, int item_layout) {
         this.context = context;
         this.addresses = addresses;
         this.item_layout = item_layout;
@@ -39,7 +36,7 @@ public class AddressAdapter extends BaseAdapter {
 
     @Override
     public long getItemId(int position) {
-        return 0;
+        return position;
     }
 
     @Override
@@ -54,6 +51,7 @@ public class AddressAdapter extends BaseAdapter {
             holder.edtphone = view.findViewById(R.id.edtphone);
             holder.edtprovince = view.findViewById(R.id.edtprovince);
             holder.edtstreet = view.findViewById(R.id.edtstress);
+            holder.name = view.findViewById(R.id.addressUsername);
 
             view.setTag(holder);
         }else {
@@ -61,12 +59,13 @@ public class AddressAdapter extends BaseAdapter {
         }
         //binding
         Addresses address = addresses.get(position);
+        holder.name.setText(address.getName());
         holder.txtdetailadress.setText(address.getStreet() + " , " + address.getProvince());
 
         return view;
     }
     public static class ViewHolder{
-        TextView txtdetailadress;
+        TextView txtdetailadress, name;
         EditText edtuser, edtphone, edtprovince, edtstreet;
     }
 }
