@@ -18,14 +18,19 @@ import android.widget.AdapterView;
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.AddressAdapter;
 import com.unipet7.mcommerce.databinding.FragmentAddressBinding;
+import com.unipet7.mcommerce.firebase.FireStoreClass;
 import com.unipet7.mcommerce.models.Addresses;
+import com.unipet7.mcommerce.utils.Constants;
 
 import java.util.ArrayList;
+import java.util.List;
 
 public class FragmentAddress extends Fragment {
     FragmentAddressBinding binding;
     AddressAdapter addressAdapter;
-    ArrayList<Addresses> addresses;
+    List<Addresses> addresses;
+
+    FireStoreClass fireStoreClass;
 
 
     @Override
@@ -34,26 +39,31 @@ public class FragmentAddress extends Fragment {
         // Inflate the layout for this fragment
         binding= FragmentAddressBinding.inflate(inflater, container, false);
         setActionBar(binding.toolbaraddress);
-        addEvents();
-        addEvents1();
-        initData();
-        loadData();
-        return binding.getRoot();
-
-    }
-    private void initData() {
         addresses = new ArrayList<>();
-        addresses.add(new Addresses("Heineken","0856433569", "Binh Dinh", "65 Nguyen Chi Thanh"));
-    }
-    private void loadData() {
-        addressAdapter = new AddressAdapter(requireContext(), addresses,R.layout.layout_address_list);
-        binding.lvlAddress.setAdapter(addressAdapter);
+        fireStoreClass = new FireStoreClass();
+        fireStoreClass.GetUserAddresses(this);
+        addEvents();
+        return binding.getRoot();
     }
 
-    private void addEvents1() {
+
+
+    public void setActionBar(@Nullable Toolbar toolbar) {
+        ((AppCompatActivity) getActivity()).setSupportActionBar(toolbar);
+        ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
+        assert actionBar != null;
+        actionBar.setDisplayHomeAsUpEnabled(true);
+        actionBar.setHomeAsUpIndicator(R.drawable.ic_back_profile);
+        actionBar.setDisplayShowTitleEnabled(false);
+    }
+    private void addEvents() {
+        binding.toolbaraddress.setNavigationOnClickListener(v -> requireActivity().getOnBackPressedDispatcher().onBackPressed());
         binding.lvlAddress.setOnItemClickListener((parent, view, position, id) -> {
             Fragment editAdrressFragment = new FragmentAddressEdit();
             FragmentTransaction transaction = getParentFragmentManager().beginTransaction();
+            Bundle bundle = new Bundle();
+            bundle.putParcelable(Constants.ADDRESS, addresses.get(position));
+            editAdrressFragment.setArguments(bundle);
             transaction.replace(R.id.fragment_container, editAdrressFragment);
             transaction.addToBackStack(null); // Thêm Fragment hiện tại vào back stack
             transaction.commit();
@@ -68,30 +78,11 @@ public class FragmentAddress extends Fragment {
     }
 
 
-    public void setActionBar(@Nullable Toolbar toolbar) {
-        ((AppCompatActivity) getActivity()).setSupportActionBar(toolbar);
-        ActionBar actionBar = ((AppCompatActivity) getActivity()).getSupportActionBar();
-        assert actionBar != null;
-        actionBar.setDisplayHomeAsUpEnabled(true);
-        actionBar.setHomeAsUpIndicator(R.drawable.ic_back_profile);
-        actionBar.setDisplayShowTitleEnabled(false);
+    public void loadAddresses(List<Addresses> addresses) {
+        if (addresses != null) {
+            this.addresses.addAll(addresses);
+            addressAdapter = new AddressAdapter(requireContext(), addresses,R.layout.layout_address_list);
+            binding.lvlAddress.setAdapter(addressAdapter);
+        }
     }
-    private void addEvents() {
-        binding.toolbaraddress.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                requireActivity().getOnBackPressedDispatcher().onBackPressed();
-            }
-        });
-//        binding.lvlAddress.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-//            @Override
-//            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-//                Intent intent = new Intent(requireContext(), FragmentAddressEdit.class);
-//                startActivity(intent);
-//            }
-//        });
-
-    }
-
-
 }
